change: move v1 network API to `openraft-network-v1` crate
`RaftNetwork`, the first version of the network trait, has been moved to a
separate `openraft-network-v1` crate. The core `openraft` crate now only
provides `RaftNetworkV2`. If you still want to use the chunk-based snapshot
transport via `RaftNetwork`, add the new crate and use the provided adapters.

Changes:
- Create `openraft-network-v1` crate with `Adapter`, `ChunkedRaft`
- Add `Adapter<C, N>` to convert `RaftNetwork` implementations to `RaftNetworkV2`
- Add `ChunkedRaft<C>` wrapper providing `install_snapshot()` for receiving chunks
- Remove `snapshot_transport.rs` and `v2/adapt_v1.rs` from openraft
- Add deprecated `RaftNetwork` stub in openraft with migration guidance
- Remove unused `tokio` dependency from openraft
- Update docs and examples to use `openraft-network-v1`

Upgrade tip:

If you are already using `RaftNetworkV2`, no changes are needed.

If you are using `RaftNetwork` (v1), follow these steps:

Add `openraft-network-v1` to your Cargo.toml:
```toml
[dependencies]
openraft = "0.10"
openraft-network-v1 = "0.10"
```

Update imports:
```rust
// Before
use openraft::RaftNetwork;

// After
use openraft_network_v1::RaftNetwork;
```

Client-side (Leader sending snapshots):
Wrap your `RaftNetwork` impl with `Adapter` to get `RaftNetworkV2`:
```rust
use openraft_network_v1::Adapter;

// Your existing RaftNetwork implementation
impl RaftNetwork<MyConfig> for MyNetwork { ... }

// In RaftNetworkFactory, wrap it with Adapter:
impl RaftNetworkFactory<MyConfig> for MyFactory {
    type Network = Adapter<MyConfig, MyNetwork>;

    async fn new_client(&mut self, ...) -> Self::Network {
        Adapter::new(MyNetwork::new(...))
    }
}
```

Server-side (Follower receiving snapshots):
Wrap your `Raft` instance with `ChunkedRaft` to handle `InstallSnapshotRequest`:
```rust
use openraft_network_v1::ChunkedRaft;

let raft = openraft::Raft::new(...).await?;
let chunked_raft = ChunkedRaft::new(raft.clone());

// In your RPC handler for InstallSnapshotRequest:
async fn handle_install_snapshot(
    chunked_raft: ChunkedRaft<MyConfig>,
    req: InstallSnapshotRequest<MyConfig>,
) -> Result<InstallSnapshotResponse<MyConfig>, _> {
    chunked_raft.install_snapshot(req).await
}

// ChunkedRaft implements Deref<Target=Raft>, so all Raft methods are available:
chunked_raft.client_write(...).await?;
chunked_raft.ensure_linearizable().await?;
```

For a complete example using `RaftNetwork` v1, see:
- `examples/raft-kv-memstore/` - full application using v1 network
- `examples/network-v1-http/` - reusable HTTP-based v1 network implementation
