+++ benchmarks/minimal/Cargo.toml
@@ -17,7 +17,8 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/databendlabs/openraft"
 
 [dependencies]
-openraft = { path = "../../openraft", version = "0.10.0", features = ["serde", "type-alias", "runtime-stats"] }
+openraft            = { path = "../../openraft", version = "0.10.0", features = ["serde", "type-alias", "runtime-stats"] }
+openraft-network-v1 = { path = "../../network-v1", version = "0.10.0" }
 
 anyhow     = { version = "1.0.63" }
 clap       = { version = "4.2", features = ["derive"] }
