+++ examples/network-v1-http/Cargo.toml
@@ -16,6 +16,7 @@ repository = "https://github.com/databendlabs/openraft"
 
 [dependencies]
 openraft = { path = "../../openraft", features = ["serde", "type-alias"] }
+openraft-network-v1 = { path = "../../network-v1" }
 
 reqwest    = { version = "0.12.5", features = ["json"] }
 serde      = { version = "1.0.114", features = ["derive"] }
