+++ examples/raft-kv-memstore-single-threaded/Cargo.toml
@@ -16,7 +16,8 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/databendlabs/openraft"
 
 [dependencies]
-openraft = { path = "../../openraft", features = ["serde", "single-threaded", "type-alias"] }
+openraft           = { path = "../../openraft", features = ["serde", "single-threaded", "type-alias"] }
+openraft-network-v1 = { path = "../../network-v1" }
 
 futures            = { version = "0.3" }
 serde              = { version = "1.0.114", features = ["derive"] }
