+++ examples/raft-kv-memstore/src/app.rs
@@ -1,14 +1,16 @@
 use std::sync::Arc;
 
+use openraft_network_v1::ChunkedRaft;
+
 use crate::NodeId;
-use crate::Raft;
 use crate::StateMachineStore;
+use crate::TypeConfig;
 
 // Representation of an application state. This struct can be shared around to share
 // instances of raft, store and more.
 pub struct App {
     pub id: NodeId,
     pub addr: String,
-    pub raft: Raft,
+    pub raft: ChunkedRaft<TypeConfig>,
     pub state_machine_store: Arc<StateMachineStore>,
 }
