+++ examples/raft-kv-rocksdb/src/app.rs
@@ -3,16 +3,17 @@ use std::sync::Arc;
 
 use futures::lock::Mutex;
 use openraft::Config;
+use openraft_network_v1::ChunkedRaft;
 
 use crate::NodeId;
-use crate::typ::Raft;
+use crate::TypeConfig;
 
 // Representation of an application state. This struct can be shared around to share
 // instances of raft, store and more.
 pub struct App {
     pub id: NodeId,
     pub addr: String,
-    pub raft: Raft,
+    pub raft: ChunkedRaft<TypeConfig>,
     pub key_values: Arc<Mutex<BTreeMap<String, String>>>,
     pub config: Arc<Config>,
 }
