+++ network-v1/src/network.rs
@@ -1,30 +1,37 @@
+//! RaftNetwork trait for chunk-based snapshot transmission.
+//!
+//! This is the v1 network API that uses `install_snapshot()` for chunked snapshot transfer.
+//! For new implementations, prefer `RaftNetworkV2` which uses `full_snapshot()`.
+
 use std::time::Duration;
 
+use openraft::OptionalSend;
+use openraft::OptionalSync;
+use openraft::RaftTypeConfig;
+use openraft::error::InstallSnapshotError;
+use openraft::error::RPCError;
+use openraft::error::RaftError;
+use openraft::network::Backoff;
+use openraft::network::RPCOption;
+use openraft::raft::AppendEntriesRequest;
+use openraft::raft::AppendEntriesResponse;
+use openraft::raft::InstallSnapshotRequest;
+use openraft::raft::InstallSnapshotResponse;
+use openraft::raft::VoteRequest;
+use openraft::raft::VoteResponse;
 use openraft_macros::add_async_trait;
-
-use crate::OptionalSend;
-use crate::OptionalSync;
-use crate::RaftTypeConfig;
-use crate::error::RPCError;
-use crate::error::RaftError;
-use crate::network::Backoff;
-use crate::network::rpc_option::RPCOption;
-use crate::raft::AppendEntriesRequest;
-use crate::raft::AppendEntriesResponse;
-use crate::raft::VoteRequest;
-use crate::raft::VoteResponse;
 
 /// A trait defining the interface for a Raft network between cluster members.
 ///
-/// See the [network chapter of the guide](crate::docs::getting_started#4-implement-raftnetwork)
-/// for details and discussion on this trait and how to implement it.
+/// This is the v1 network API that uses chunk-based snapshot transmission via
+/// `install_snapshot()`. For new implementations, consider using
+/// [`RaftNetworkV2`](openraft::network::v2::RaftNetworkV2) instead.
+///
+/// To use this trait with openraft, wrap your implementation with
+/// [`Adapter`](crate::Adapter) to convert it to `RaftNetworkV2`.
 ///
 /// A single network instance is used to connect to a single target node. The network instance is
-/// constructed by the [`RaftNetworkFactory`](`crate::network::RaftNetworkFactory`).
-///
-/// [Ensure connection to correct node][correct-node]
-///
-/// [correct-node]: `crate::docs::cluster_control::dynamic_membership#ensure-connection-to-the-correct-node`
+/// constructed by the [`RaftNetworkFactory`](`openraft::network::RaftNetworkFactory`).
 #[add_async_trait]
 pub trait RaftNetwork<C>: OptionalSend + OptionalSync + 'static
 where C: RaftTypeConfig
@@ -45,9 +52,9 @@ where C: RaftTypeConfig
     /// Called when a follower is too far behind to use log replication.
     async fn install_snapshot(
         &mut self,
-        _rpc: crate::raft::InstallSnapshotRequest<C>,
-        _option: RPCOption,
-    ) -> Result<crate::raft::InstallSnapshotResponse<C>, RPCError<C, RaftError<C, crate::error::InstallSnapshotError>>>;
+        rpc: InstallSnapshotRequest<C>,
+        option: RPCOption,
+    ) -> Result<InstallSnapshotResponse<C>, RPCError<C, RaftError<C, InstallSnapshotError>>>;
 
     /// Send a RequestVote RPC to the target.
     ///
@@ -61,7 +68,7 @@ where C: RaftTypeConfig
 
     /// Build a backoff instance if the target node is temporarily(or permanently) unreachable.
     ///
-    /// When a [`Unreachable`](`crate::error::Unreachable`) error is returned from the `Network`
+    /// When a [`Unreachable`](`openraft::error::Unreachable`) error is returned from the `Network`
     /// methods, Openraft does not retry connecting to a node immediately. Instead, it sleeps
     /// for a while and retries. The duration of the sleep is determined by the backoff
     /// instance.
