+++ openraft/src/docs/faq/07-troubleshooting/06-excessive-network-errors.md
@@ -5,10 +5,10 @@ when a follower is unreachable
 
 **Cause**: Openraft retries replication aggressively. Each failed RPC logs an error.
 
-**Solution**: In your [`RaftNetwork`][] implementation, when a node is known to be unreachable,
+**Solution**: In your [`RaftNetworkV2`][] implementation, when a node is known to be unreachable,
 return [`Unreachable`][] error instead of [`NetworkError`][]. Openraft backs off longer for
 `Unreachable` errors, reducing log spam.
 
-[`RaftNetwork`]: `crate::network::RaftNetwork`
+[`RaftNetworkV2`]: `crate::network::RaftNetworkV2`
 [`Unreachable`]: `crate::error::Unreachable`
 [`NetworkError`]: `crate::error::NetworkError`
