+++ openraft/src/docs/faq/faq.md
@@ -375,9 +375,9 @@ a list of active nodes without modifying cluster membership.
 
 ### How to minimize error logging when a follower is offline
 
-Excessive error logging, like `ERROR openraft::replication: 248: RPCError err=NetworkError: ...`, occurs when a follower node becomes unresponsive. To alleviate this, implement a mechanism within [`RaftNetwork`][] that returns a [`Unreachable`][] error instead of a [`NetworkError`][] when immediate replication retries to the affected node are not advised.
+Excessive error logging, like `ERROR openraft::replication: 248: RPCError err=NetworkError: ...`, occurs when a follower node becomes unresponsive. To alleviate this, implement a mechanism within [`RaftNetworkV2`][] that returns a [`Unreachable`][] error instead of a [`NetworkError`][] when immediate replication retries to the affected node are not advised.
 
-[`RaftNetwork`]: `crate::network::RaftNetwork`
+[`RaftNetworkV2`]: `crate::network::RaftNetworkV2`
 [`Unreachable`]: `crate::error::Unreachable`
 [`NetworkError`]: `crate::error::NetworkError`
 
@@ -724,11 +724,11 @@ when a follower is unreachable
 
 **Cause**: Openraft retries replication aggressively. Each failed RPC logs an error.
 
-**Solution**: In your [`RaftNetwork`][] implementation, when a node is known to be unreachable,
+**Solution**: In your [`RaftNetworkV2`][] implementation, when a node is known to be unreachable,
 return [`Unreachable`][] error instead of [`NetworkError`][]. Openraft backs off longer for
 `Unreachable` errors, reducing log spam.
 
-[`RaftNetwork`]: `crate::network::RaftNetwork`
+[`RaftNetworkV2`]: `crate::network::RaftNetworkV2`
 [`Unreachable`]: `crate::error::Unreachable`
 [`NetworkError`]: `crate::error::NetworkError`
