+++ openraft/src/docs/upgrade_guide/upgrade-v08-v09.md
@@ -181,15 +181,15 @@ To use arbitrary snapshot data, the application needs to:
 
 [`AsyncRuntime`]:                     `crate::AsyncRuntime`
 [`RPCOption`]:                        `crate::network::RPCOption`
-[`Chunked`]:                          `crate::network::snapshot_transport::Chunked`
-[`Streaming`]:                        `crate::network::snapshot_transport::Streaming`
+[`Chunked`]:                          https://docs.rs/openraft-network-v1/latest/openraft_network_v1/struct.Chunked.html
+[`Streaming`]:                        https://docs.rs/openraft-network-v1/latest/openraft_network_v1/struct.Streaming.html
 [`Raft::ensure_linearizable()`]:      `crate::Raft::ensure_linearizable`
 [`Raft::get_snapshot()`]:             `crate::Raft::get_snapshot`
 [`Raft::begin_receiving_snapshot()`]: `crate::Raft::begin_receiving_snapshot`
 [`Raft::install_full_snapshot()`]:    `crate::Raft::install_full_snapshot`
 
 [`RaftNetwork`]:                      `crate::network::RaftNetwork`
-[`RaftNetwork::full_snapshot()`]:     `crate::network::v2::RaftNetworkV2::full_snapshot`
+[`RaftNetwork::full_snapshot()`]:     `crate::network::RaftNetworkV2::full_snapshot`
 
 [`RaftLogStorage::save_committed()`]: `crate::storage::RaftLogStorage::save_committed`
