--- openraft/src/network/v1/mod.rs
@@ -1,7 +1,0 @@
-//! Version 1 of the Raft network API.
-
-mod factory;
-mod network;
-
-pub use factory::RaftNetworkFactory;
-pub use network::RaftNetwork;
