+++ openraft/src/raft/message/append_entries_response.rs
@@ -9,11 +9,11 @@ use crate::type_config::alias::VoteOf;
 
 /// The response to an `AppendEntriesRequest`.
 ///
-/// [`RaftNetwork::append_entries`] returns this type only when received an RPC reply.
+/// [`RaftNetworkV2::append_entries`] returns this type only when received an RPC reply.
 /// Otherwise, it should return [`RPCError`].
 ///
 /// [`RPCError`]: crate::error::RPCError
-/// [`RaftNetwork::append_entries`]: crate::network::RaftNetwork::append_entries
+/// [`RaftNetworkV2::append_entries`]: crate::network::RaftNetworkV2::append_entries
 #[derive(Debug, Clone)]
 #[derive(PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize), serde(bound = ""))]
@@ -23,7 +23,7 @@ pub enum AppendEntriesResponse<C: RaftTypeConfig> {
 
     /// Successfully sent the first portion of log entries.
     ///
-    /// [`RaftNetwork::append_entries`] can return a partial success.
+    /// [`RaftNetworkV2::append_entries`] can return a partial success.
     /// For example, it tries to send log entries `[1-2..3-10]`, the application is allowed to send
     /// just `[1-2..1-3]` and return `PartialSuccess(1-3)`
     ///
@@ -31,12 +31,12 @@ pub enum AppendEntriesResponse<C: RaftTypeConfig> {
     ///
     /// The returned matching log id must be **greater than or equal to** the first log
     /// id([`AppendEntriesRequest::prev_log_id`]) of the entries to send. If no RPC reply is
-    /// received, [`RaftNetwork::append_entries`] must return an [`RPCError`] to inform
+    /// received, [`RaftNetworkV2::append_entries`] must return an [`RPCError`] to inform
     /// Openraft that the first log id([`AppendEntriesRequest::prev_log_id`]) may not match on
     /// the remote target node.
     ///
     /// [`RPCError`]: crate::error::RPCError
-    /// [`RaftNetwork::append_entries`]: crate::network::RaftNetwork::append_entries
+    /// [`RaftNetworkV2::append_entries`]: crate::network::RaftNetworkV2::append_entries
     /// [`AppendEntriesRequest::prev_log_id`]: crate::raft::AppendEntriesRequest::prev_log_id
     PartialSuccess(Option<LogIdOf<C>>),
