+++ openraft/src/raft/mod.rs
@@ -890,59 +890,6 @@ where C: RaftTypeConfig
         self.protocol_api().install_full_snapshot(vote, snapshot).await
     }
 
-    /// Receive an `InstallSnapshotRequest`.
-    ///
-    /// These RPCs are sent by the cluster leader in order to bring a new node or a slow node
-    /// up-to-speed with the leader.
-    ///
-    /// If receiving is finished `done == true`, it installs the snapshot to the state machine.
-    /// Nothing will be done if the input snapshot is older than the state machine.
-    #[tracing::instrument(level = "debug", skip_all)]
-    #[cfg(feature = "tokio-rt")]
-    pub async fn install_snapshot(
-        &self,
-        req: InstallSnapshotRequest<C>,
-    ) -> Result<InstallSnapshotResponse<C>, RaftError<C, crate::error::InstallSnapshotError>>
-    where
-        C::SnapshotData: tokio::io::AsyncRead + tokio::io::AsyncWrite + tokio::io::AsyncSeek + Unpin,
-    {
-        use crate::async_runtime::mutex::Mutex;
-        use crate::vote::raft_vote::RaftVoteExt;
-
-        tracing::debug!("Raft::install_snapshot(): req: {}", req);
-
-        let req_vote = req.vote.clone();
-        let my_vote = self.with_raft_state(|state| state.vote_ref().clone()).await?;
-        let resp = InstallSnapshotResponse { vote: my_vote.clone() };
-
-        // Check vote.
-        // It is not mandatory because it is just a read operation
-        // but prevent unnecessary snapshot transfer early.
-        {
-            if req_vote.as_ref_vote() >= my_vote.as_ref_vote() {
-                // Ok
-            } else {
-                tracing::info!("vote {} is rejected by local vote: {}", req_vote, my_vote);
-                return Ok(resp);
-            }
-        }
-
-        let finished_snapshot = {
-            use crate::network::snapshot_transport::Chunked;
-            use crate::network::snapshot_transport::StreamingState;
-
-            let streaming_state = self.inner.extensions.get::<StreamingState<C>>();
-            let mut streaming = streaming_state.streaming.lock().await;
-            Chunked::<C>::receive_snapshot(&mut *streaming, self, req).await?
-        };
-
-        if let Some(snapshot) = finished_snapshot {
-            let resp = self.install_full_snapshot(req_vote, snapshot).await?;
-            return Ok(resp.into());
-        }
-        Ok(resp)
-    }
-
     /// Get the ID of the current leader from this Raft node.
     ///
     /// This method is based on the Raft metrics system which does a good job at staying
@@ -1205,7 +1152,7 @@ where C: RaftTypeConfig
     /// the cluster via [`RaftNetworkV2::transfer_leader`] and the implementation on the remote node
     /// responds to transfer leader request by calling this method.
     ///
-    /// [`RaftNetworkV2::transfer_leader`]: crate::network::v2::RaftNetworkV2::transfer_leader
+    /// [`RaftNetworkV2::transfer_leader`]: crate::network::RaftNetworkV2::transfer_leader
     #[since(version = "0.10.0")]
     #[tracing::instrument(level = "debug", skip_all)]
     pub async fn handle_transfer_leader(&self, req: TransferLeaderRequest<C>) -> Result<(), Fatal<C>> {
